+++ core-ng/src/test/java/core/framework/internal/db/RepositoryImplAssignedIdEntityTest.java
@@ -171,14 +171,14 @@ class RepositoryImplAssignedIdEntityTest {
     @Test
     void batchInsertIgnore() {
         List<AssignedIdEntity> entities = Lists.newArrayList();
-        for (int i = 0; i < 10; i++) {
+        for (int i = 0; i < 5; i++) {
             AssignedIdEntity entity = entity(String.valueOf(i), "value" + i, 10 + i);
             entities.add(entity);
             entities.add(entity);
         }
-        int inserted = repository.batchInsertIgnore(entities);
+        boolean[] results = repository.batchInsertIgnore(entities);
 
-        assertThat(inserted).isEqualTo(10);
+        assertThat(results).hasSize(10).contains(true, false, true, false, true, false, true, false, true, false);
         assertThat(repository.get("0")).get().usingRecursiveComparison().isEqualTo(entities.get(0));
         assertThat(repository.get("1")).get().usingRecursiveComparison().isEqualTo(entities.get(2));
     }